commit on 08-Aug

+++ backend/src/main/java/org/sid/ebankingbackend/repositories/LoanRepository.java
@@ -1,16 +1,16 @@
 package org.sid.ebankingbackend.repositories;
-import java.util.Date;
+//import java.util.Date;
 import java.util.List;
 
-import org.sid.ebankingbackend.entities.Customer;
+//import org.sid.ebankingbackend.entities.Customer;
 import org.sid.ebankingbackend.entities.LoanEntity;
-import org.sid.ebankingbackend.entities.Payment;
+//import org.sid.ebankingbackend.entities.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.lang.NonNull;
+//import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Repository;
-import org.sid.ebankingbackend.enums.LoanStatus;
+//import org.sid.ebankingbackend.enums.LoanStatus;
 
 @Repository
 public interface LoanRepository extends JpaRepository<LoanEntity, Long> {
@@ -21,6 +21,8 @@ public interface LoanRepository extends JpaRepository<LoanEntity, Long> {
  
     @Query("select l from LoanEntity l where l.customer.id like :kw")
     List<LoanEntity> findByCustomerId(@Param("kw") Long customerId);
+
+    
  
     @Query("SELECT l FROM LoanEntity l WHERE (:loan_id IS NULL OR l.loan_id=:loan_id) AND (:plan_id IS NULL OR l.loanPlan.plan_id=:plan_id ) AND (:type_id IS NULL OR (l.loanType.type_id=:type_id) ) AND (:status IS NULL OR l.status = :status) AND ((l.loan_amt <= :max_amount AND l.loan_amt >= :min_amount ))")
     List<LoanEntity> findByAttributes(
@@ -28,8 +30,8 @@ public interface LoanRepository extends JpaRepository<LoanEntity, Long> {
             @Param("plan_id") Long plan_id,
             @Param("type_id") Long type_id,
             @Param("status") String status,
-            @Param("min_amount") Integer min_amount,
-            @Param("max_amount") Integer max_amount
+            @Param("min_amount") double min_amount,
+            @Param("max_amount") double max_amount
             );
 
 }
